revisit tcplay support

+++ zuluCrypt-cli/lib/file_path_security.c
@@ -50,19 +50,22 @@ int zuluCryptSecureOpenFile( const char * path,int * fd,string_t * file,uid_t ui
 
 void zuluCryptDeleteFile( const char * file )
 {
+	int fd ;
 	void * map ;
 	struct stat st ;
-	int fd = open( file,O_WRONLY ) ;
-	if( fd != -1 ){
-		fstat( fd,&st ) ;
-		map =  mmap( 0,st.st_size,PROT_WRITE,MAP_PRIVATE,fd,0 ) ;
-		if( map != MAP_FAILED ){
-			memset( map,'\0',st.st_size ) ;
-			munmap( map,st.st_size ) ;
+	if( file != NULL ){
+		fd = open( file,O_WRONLY ) ;
+		if( fd != -1 ){
+			fstat( fd,&st ) ;
+			map =  mmap( 0,st.st_size,PROT_WRITE,MAP_PRIVATE,fd,0 ) ;
+			if( map != MAP_FAILED ){
+				memset( map,'\0',st.st_size ) ;
+				munmap( map,st.st_size ) ;
+			}
+			close( fd ) ;
 		}
-		close( fd ) ;
+		unlink( file ) ;
 	}
-	unlink( file ) ;
 }
 
 static int _check_if_device_is_supported( int st,uid_t uid,char ** dev )
